% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autos.R
\name{attach_auto}
\alias{attach_auto}
\title{Attach a function directory}
\usage{
attach_auto(path, name)
}
\arguments{
\item{path}{Directory path}

\item{name}{Directory name}
}
\value{
Environment containing functions
}
\description{
This function is used to create an rautos path. All .R files from a given
path are sourced. Functions are imported into from a directory and returned
to an environment. This environment can then be used to attach on to the
search path. This function should not be called directly. To apply autos,
use \code{set_autos()}.
}
\examples{
\dontrun{
attach_auto("./my_funcs", "my_autos")
}
}
