% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autos.R
\name{detach_autos}
\alias{detach_autos}
\title{Detach the autos from the current session}
\usage{
detach_autos(names)
}
\arguments{
\item{names}{vector of names of attached packages to detach}
}
\value{
names found in the search path
}
\description{
This function will remove any autos that have been set from the search path
}
\examples{
\dontrun{
detach_autos()
}
}
