% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprofile.R
\name{rprofile}
\alias{rprofile}
\title{Function used to pass through code to the .Rprofile}
\usage{
rprofile(config)
}
\arguments{
\item{config}{configuration object from config::get()}
}
\value{
Directory paths of the R autos
}
\description{
Function used to pass through code to the .Rprofile
}
\examples{
\dontrun{
rprofile(config::get("path/to/config/_envsetup.yml"))
}
}
