% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autos.R
\name{set_autos}
\alias{set_autos}
\title{Set the R autos}
\usage{
set_autos(..., envsetup_environ = Sys.getenv("ENVSETUP_ENVIRON"))
}
\arguments{
\item{...}{named vector of directories}

\item{envsetup_environ}{name of the environment you would like to read from;
default values comes from the value in the system variable ENVSETUP_ENVIRON
which can be set by Sys.setenv(ENVSETUP_ENVIRON = "environment name")}
}
\value{
Directory paths of the R autos
}
\description{
Set the directory paths of any 'autos'. 'Autos' are
directory paths that hold .R files containing R functions. These paths may be
used when functions apply to an analysis, protocol, or even at a global
level, but don't fit in or necessarily require a package or haven't been
incorporated into a package.
}
\examples{
\dontrun{
set_autos(envsetup_config$autos)
}
}
