% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\docType{data}
\name{envsetup_environment}
\alias{envsetup_environment}
\title{Path environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
envsetup_environment
}
\description{
Path environment
}
\keyword{datasets}
