% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize the R environment with envsetup}
\usage{
init(project = getwd(), config_path = NULL)
}
\arguments{
\item{project}{Character. The path to the project directory. Defaults to the current working directory.}

\item{config_path}{Character. The path of the config file. Defaults to "envsetup.yml".}
}
\value{
Logical. TRUE if successful.
}
\description{
Initialize the R environment with envsetup
}
\examples{
\dontrun{
init()
}
}
