% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{read_path}
\alias{read_path}
\title{Read path}
\usage{
read_path(
  lib,
  filename,
  full.path = TRUE,
  envsetup_environ = Sys.getenv("ENVSETUP_ENVIRON")
)
}
\arguments{
\item{lib}{object containing the paths for all environments of a directory}

\item{filename}{name of the file you would like to read}

\item{full.path}{logical to return the path including the file name}

\item{envsetup_environ}{name of the environment you would like to read the file from;
default values comes from the value in the system variable ENVSETUP_ENVIRON
which can be set by Sys.setenv(ENVSETUP_ENVIRON = "environment name")}
}
\value{
path of the first place the file is found
}
\description{
Check each environment for the file and return the path to the first.
}
\details{
The environments searched depends on the current environment.
For example, if your workflow contains a development (dev) area and
production area (prod), and the code is executing in the dev environment,
we search dev and prod. If in prod, we only search prod.
}
\examples{
\dontrun{
read_path(a_in, "adsl.sas7bdat")
}
}
