% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{write_path}
\alias{write_path}
\title{Write path}
\usage{
write_path(lib, envsetup_environ = Sys.getenv("ENVSETUP_ENVIRON"))
}
\arguments{
\item{lib}{object containing the paths for all environments of a directory}

\item{envsetup_environ}{name of the environment you would like to write to}
}
\value{
path to write
}
\description{
Write path
}
\examples{
\dontrun{
write_path(a_in, "PROD")
}
}
