% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{write_path}
\alias{write_path}
\title{Retrieve a file path from an envsetup object containing paths}
\usage{
write_path(
  lib,
  filename = NULL,
  envsetup_environ = Sys.getenv("ENVSETUP_ENVIRON")
)
}
\arguments{
\item{lib}{Object containing the paths for all environments of a directory}

\item{filename}{Name of the file you would like to write}

\item{envsetup_environ}{Name of the environment to which you would like to
write. Defaults to the ENVSETUP_ENVIRON environment variable}
}
\value{
path to write
}
\description{
Paths will be filtered to produce the lowest available level from a hierarchy
of paths based on envsetup_environ
}
\examples{
\dontrun{
write_path(a_in, "PROD")
}
}
